Add support for animation-range with ScrollTimeline
